pyproject.toml
[project]
name = "binsim-analyzer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "pyyaml",
    "pyelftools",
    "pylzma",
    "msgpack",
    "pyghidra",
    "ghidra-stubs==12.0.4"
]
